Milvus POC/build_rag/requirements.txt
pymilvus
openai
requests
tqdm
python-dotenv

